fix: make first-import order independent around the workshop package
kai.workshop's package initialization eagerly imports the execution
stack, which reaches back into top-level modules; two import rings only
survived because production happens to import kai.workshop first.
Importing kai.conversation_compatibility first failed (it pulled the
workshop package, whose chain re-imported it mid-initialization), and
importing kai.pool first failed the same way through kai.sessions and
kai.workshop.runtime_pool.

- CanonicalMemoryProvenance moves to kai.workshop.domain, beside its
  field types, in the package's only kai-import-free module. With it
  gone, kai.conversation_compatibility no longer imports kai.workshop
  at runtime at all (a type-only import remains for its signature).
- kai.workshop.runtime_pool and kai.workshop.protected_execution use
  kai.pool names purely as annotations, so those imports become
  type-only.
- kai.pool defers its workshop exception import to the one method that
  catches it, removing its last module-level workshop dependency.

A new test imports each cycle-prone entry module in a fresh interpreter,
where it must load as the process's first kai import; in-process imports
would inherit sys.modules state from earlier tests and prove nothing.
